requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
modal==0.63.0
google-api-python-client==2.100.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
twilio==8.10.0
pandas==2.1.3
pydantic==2.5.0
requests==2.31.0
langchain>=0.1.0
langchain-openai>=0.1.0
langchainhub>=0.1.14

